More features (unfinished)
Many additions. We have managed to draw all the logos of the characters for the screen of selection of fighters for the teams. And we have also drawn the information screen of, so far, only the attack class. Now in this version, when you right click on any of the logos, a page with the character's appearance, a small description of the character's profile and obviously his stats will be shown on the screen. The interesting thing about the stats is that the character's stats are not already drawn in the design, but will be dynamically rendered from the loaded Fighter file of the character, this is done by the function “showinfoandstats()”.