Add files via upload
In this commit I have managed to solve the problem of the crashes and the problem that caused the second team to have the same fighters as team 1. 
The crashes problem occurred during the SDL_Delay that allowed to have the message “You only can pick 4” for 2 seconds on screen. During the display of this message if you tried to quickly click on other logos to select more fighters in the team ,  the clicks would accumulate in the SDL_PollEvent queue at the start of the while loop. The fact that each click processed caused the use of SDL_Delay(2000) which is a relatively slow and heavy to process  this led to the game crashing. The solution I have done is to add a while loop that does its own SDL_PollEvent inside instead of the SDL_Delay directly, this loop allows to pass all the clicks making almost impossible to accumulate enough clicks to crash the game.